backend/requirements.txt
dotenv

django
djangorestframework
psycopg2-binary
drf-yasg
django-filter